updated .toggle_redshift.sh to support restoring the current state because calling xcalib overrides the redshift setting, which was a problem upon resuming

+++ X/.toggle_redshift.sh
@@ -3,6 +3,8 @@ DIM1_INDICATOR=$HOME/.toggle_redshift_indicator_dim1
 DIM2_INDICATOR=$HOME/.toggle_redshift_indicator_dim2
 DIM3_INDICATOR=$HOME/.toggle_redshift_indicator_dim3
 
+DEFAULT_COLOR=$HOME/.xcalib.sh
+
 DIM1_GAMMA='1:1:1'
 DIM1_TEMP=5000
 
@@ -12,21 +14,36 @@ DIM2_TEMP=4200
 DIM3_GAMMA='1:1:1'
 DIM3_TEMP=3500
 
-if [ ! -e $DIM1_INDICATOR ] && [ ! -e $DIM2_INDICATOR ] && [ ! -e $DIM3_INDICATOR ]; then
-  touch $DIM1_INDICATOR
-  redshift -P -g $DIM1_GAMMA -O $DIM1_TEMP
+if [ -n "$1" ] && [ "$1" = "restore" ]; then
+  if [ ! -e $DIM1_INDICATOR ] && [ ! -e $DIM2_INDICATOR ] && [ ! -e $DIM3_INDICATOR ]; then
+    . $DEFAULT_COLOR
 
-elif [ -e $DIM1_INDICATOR ]; then
-  rm $DIM1_INDICATOR
-  touch $DIM2_INDICATOR
-  redshift -P -g $DIM2_GAMMA -O $DIM2_TEMP
+  elif [ -e $DIM1_INDICATOR ]; then
+    redshift -P -g $DIM1_GAMMA -O $DIM1_TEMP
 
-elif [ -e $DIM2_INDICATOR ]; then
-  rm $DIM2_INDICATOR
-  touch $DIM3_INDICATOR
-  redshift -P -g $DIM3_GAMMA -O $DIM3_TEMP
+  elif [ -e $DIM2_INDICATOR ]; then
+    redshift -P -g $DIM2_GAMMA -O $DIM2_TEMP
 
-elif [ -e $DIM3_INDICATOR ]; then
-  rm $DIM3_INDICATOR
-  redshift -x
+  elif [ -e $DIM3_INDICATOR ]; then
+    redshift -P -g $DIM3_GAMMA -O $DIM3_TEMP
+  fi
+else
+  if [ ! -e $DIM1_INDICATOR ] && [ ! -e $DIM2_INDICATOR ] && [ ! -e $DIM3_INDICATOR ]; then
+    touch $DIM1_INDICATOR
+    redshift -P -g $DIM1_GAMMA -O $DIM1_TEMP
+
+  elif [ -e $DIM1_INDICATOR ]; then
+    rm $DIM1_INDICATOR
+    touch $DIM2_INDICATOR
+    redshift -P -g $DIM2_GAMMA -O $DIM2_TEMP
+
+  elif [ -e $DIM2_INDICATOR ]; then
+    rm $DIM2_INDICATOR
+    touch $DIM3_INDICATOR
+    redshift -P -g $DIM3_GAMMA -O $DIM3_TEMP
+
+  elif [ -e $DIM3_INDICATOR ]; then
+    rm $DIM3_INDICATOR
+    . $DEFAULT_COLOR
+  fi
 fi
